added insertion

+++ programming_paradigm/robust_division_calculator.py
@@ -1,6 +1,6 @@
 def safe_divide(numerator, denominator):
-        numerator = int(input("Enter a number: "))
-        denominator = int(input("Enter a number: "))
+        numerator = float(input("Enter a number: "))
+        denominator = float(input("Enter a number: "))
         result = numerator / denominator
         print(f"The result of the division is {result:.1f}")
 try:
